+++ logger.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <string>
 
 namespace logger {
 enum class LOG_LEVEL {
