+++ pre_ast/CMakeLists.txt
@@ -2,4 +2,10 @@ cmake_policy(SET CMP0076 NEW)
 target_sources(cornpiler 
 PUBLIC 
 ast.cpp
+filereader.cpp
+scopes.cpp
 )
+
+add_subdirectory(with)
+add_subdirectory(types)
+add_subdirectory(nodes)
