+++ pre_ast/filereader.hpp
@@ -1,6 +1,6 @@
 #pragma once
+#include "../logger.hpp"
 #include <string>
-
 class file_object {
  public:
   std::string contents;
@@ -8,3 +8,4 @@ class file_object {
 
   file_object(std::string c, int l) : contents(c), length(l){};
 };
+file_object read_file(const char *filename, logger::logger *logger);
