+++ pre_ast/nodes/const_int.cpp
@@ -1,0 +1,6 @@
+#include "const_int.hpp"
+
+ast_types::const_int::const_int() { act = act_type::const_int; }
+std::string ast_types::const_int::print_node() {
+  return "\"const_int\": {\"value\":" + this->value.print_node() + "}";
+}
