+++ pre_ast/nodes/expr.cpp
@@ -1,0 +1,6 @@
+#include "expr.hpp"
+
+ast_types::expr::expr() { act = act_type::expr; }
+std::string ast_types::expr::print_node() {
+  return "\"expr\": {\"args\": " + this->args.print_node() + "}";
+}
